fix(i18n): card lang/hreflang, feed filter on home, og:locale/hreflang, oxuduğum dillər
- Per-card language: PostCard.svelte:50 article lang={post.language},
  Button hreflang={post.language} so each card advertises its language;
  article page [handle]/[slug]/+page.svelte:254 <article lang> inherits to
  body. Fixes invisible AZ/EN/PL/IT mixing where screen readers and
  search engines saw every card as English.

- Feed language filter: extract FeedLanguageFilter.svelte (show/hide mode +
  chips + add language, bilingual title Oxuduğum dillər · Feed languages)
  reused in Settings and on home feed (+page.svelte:178 filter card above
  tabs). Home now shows the filter inline so readers can instantly hide
  unwanted languages instead of hunting in Settings; $effect watches
  reading.feedLangs/mode and refetches Local/Global when changed.
  Settings section replaced with component, label bilingual.

- hreflang / og:locale signals: +layout.svelte:283 emits
  <meta property="og:locale" content="az|en…"> plus
  <link rel="alternate" hreflang="az" href="{canonical}"> and
  x-default for every page; post pages use post.language, others fall
  back to x-default/en. Complements hooks.server's <html lang> (post
  language → page language) and card-level lang/hreflang for crawlers.
