+++ apps/frontend/src/routes/[handle]/[slug]/+page.svelte
@@ -251,7 +251,7 @@
 
 <PageTitle text={post.title ?? "Post"} />
 
-<article>
+<article lang={post.language ?? undefined}>
   {#if deleteError}
     <p class="mb-6 rounded-input border border-destructive/30 bg-destructive/10 px-4 py-3 text-sm text-destructive">
       {deleteError}
